requirements.txt
python-dotenv~=1.0.1
Flask~=3.1.0
SQLAlchemy~=2.0.36
requests~=2.32.3
pandas~=2.2.3
numpy~=2.2.1
pytz~=2024.2
Levenshtein~=0.26.1